feat: Add multilanguage support for LavaMusic
This commit adds multilanguage support to LavaMusic by providing a guide on how to add a new language. It includes instructions on creating a new file in the `locales` directory with the language code, copying the contents of the `EnglishUS.json` file, and translating the strings to the desired language. It also provides an example of the translation JSON format and formatting tags for the i18n NPM package.